% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeColname.R
\name{ChangeColname}
\alias{ChangeColname}
\title{ChangeColname}
\usage{
ChangeColname(dataset, old_name, new_name)
}
\arguments{
\item{dataset}{Dataset}

\item{old_name}{Column name of the column that you want to change}

\item{new_name}{Name for the new column}
}
\value{
dataset
}
\description{
ChangeColname
}
\examples{

ChangeColname(LandBird_Monitoring, "Condition", "new_condition")
}
