% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datatype.R
\name{datatype}
\alias{datatype}
\title{Change the datatype}
\usage{
datatype(dataset, colname, type)
}
\arguments{
\item{dataset}{Dataset}

\item{colname}{The name of the column of the data that needs to be changed}

\item{type}{The type of data you want the column to change in}
}
\value{
Dataset with the column that has the changed datatype
}
\description{
Change the datatype
}
\examples{
datatype(LandBird_Monitoring, "Year", "integer")
}
