% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer.R
\name{integer}
\alias{integer}
\title{Change datatype to integer}
\usage{
integer(data, colname)
}
\arguments{
\item{data}{Dataset}

\item{colname}{Column from dataset that needs datatype to be changed}
}
\value{
Dataset with the new column datatype
}
\description{
Change datatype to integer
}
\examples{
integer(LandBird_Monitoring, "Year")
}
