% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandBird_Monitoring.R
\docType{data}
\name{LandBird_Monitoring}
\alias{LandBird_Monitoring}
\title{LandBird_Monitoring dataset}
\format{
An object of class \code{NULL} of length 0.
}
\source{
\url{https://catalog.data.gov/dataset/ncrn-bird-monitoring-data-2007-2017}
}
\usage{
data("LandBird_Monitoring")
}
\description{
This dataset contains cleaned and processed data on bird monitoring from 2007 to 2017.
}
\examples{
data("LandBird_Monitoring")
}
\keyword{datasets}
