% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{Normalization}
\alias{Normalization}
\title{Title}
\usage{
Normalization(dataset, colname, value, data_col)
}
\arguments{
\item{dataset}{Dataset}

\item{colname}{Column name of the value that you want to normalize}

\item{value}{The value that needs to be normalized}

\item{data_col}{The data that the normalization is used on}
}
\value{
dataset
}
\description{
Title
}
\examples{

Normalization(LandBird_Monitoring, "Condition", "Calm", "Humidity")
}
