% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotZero.R
\name{NotZero}
\alias{NotZero}
\title{Notzero}
\usage{
NotZero(dataset, colname)
}
\arguments{
\item{dataset}{Dataset you want to use}

\item{colname}{The column you want to add the 0.001 to}
}
\value{
dataset
}
\description{
Notzero
}
\examples{
LandBird_Monitoring$Year <- as.numeric(LandBird_Monitoring$Year)
NotZero(LandBird_Monitoring, "Year")
}
