% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric.R
\name{numeric}
\alias{numeric}
\title{Change datatype to numeric}
\usage{
numeric(data, colname)
}
\arguments{
\item{data}{Dataset with a column}

\item{colname}{Column from the dataset that needs to be changes}
}
\value{
Dataset with the new column datatype
}
\description{
Change datatype to numeric
}
\examples{
numeric(LandBird_Monitoring, "Year")
}
