% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveNA.R
\name{RemoveNA}
\alias{RemoveNA}
\title{RemoveNA}
\usage{
RemoveNA(dataset)
}
\arguments{
\item{dataset}{The dataset you want to use}

\item{colname}{The column of the dataset where you want the NA values to be gone}
}
\value{
dataset
}
\description{
RemoveNA
}
\examples{
RemoveNA(LandBird_Monitoring, "Site_Name")
}
