% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{factor}
\alias{factor}
\title{Change datatype to a factor}
\usage{
factor(data, colname)
}
\arguments{
\item{data}{Dataset with a column}

\item{colname}{Column from the dataset that needs to be changes}
}
\value{
Dataset with the new column datatype
}
\description{
Change datatype to a factor
}
\examples{
factor(LandBird_Monitoring, "Year")
}
