% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandBird_Monitoring.R
\docType{data}
\name{LandBird_Monitoring}
\alias{LandBird_Monitoring}
\title{LandBird_Monitoring dataset}
\format{
Een data.frame met X rijen en Y kolommen
}
\source{
\url{https://catalog.data.gov/dataset/ncrn-bird-monitoring-data-2007-2017}
}
\usage{
data("LandBird_Monitoring")
}
\description{
LandBird_Monitoring dataset
}
\examples{
data("LandBird_Monitoring")
head(LandBird_Monitoring)
}
\keyword{datasets}
